Login and Logout
A "middleware" function was used in the route file for login operations.
A simple "login" function was added to the controller.
For logging out, a route was added to the route file.
The controller and function were created in the added route.
The interface was edited so that the operations could be executed by the user.